fix: create parent recording directories if necessary
This app expects the file name to be the base path - ie all files go into
the root of the recording folder specified in the config. That makes
things a bit difficult because BBB records files in different
hierarchies, so parent dirs up to recorder.directory need to be flexible
somehow. Those extra directories are created on demand based on
server-provided data generated on the fly (basically the meetingID).
An example of that flexible structure:
  - For cameras, recordings should be in
    /var/lib/bbb-webrtc-recorder/recordings/<meetingID>/<fileName>.webm
  - For screens, recordings should be in
    /var/lib/bbb-webrtc-recorder/screenshare/<meetingID><fileName>.webm

This commit checks whether the parent directories up recorder.directory
are present and try to create them if they aren't. It errors out in case
that fails.