feat: extend RPCs to support different recording adapters
Extend the startRecording RPC so that it support different "recording
adapters". A recording adapter is supposed to reflect the media server
that the recorder is pulling media from.

The idea is to support two adapters, initially: LiveKit and mediasoup.
While mediasoup is fully implemented, retaining the implementation we
had in v0.8.0, the LiveKit implementation will come in subsequent commits.

Additionally, add better RPC validation and validation tests as well.