fix: retry Redis reconn if it drops
Add an exp backoff retry to the Redis PubSub module when subscriptions
drop, but it had previously connected successfully beforehand. This is
to prevent losing recordings due to a temporary disconn.
If it fails to connect on boot, exit with a failure core and properly
log it.

Things to do in the future:
  - Send a status event on a Redis reconnection
  - Add current active recording tracks to status events for resync in
    clients