fix(livekit): add stop-and-notify handling on LK participant disconns
Handle LK participant disconnections in the simplest way we can: close
down the connection and notify clients. Retries are on the client's
account, so it's best if we keep this simple.

Additionally, unify connection states between adapters so that they're
the same and we can use the same conn state callback signature in them.