fix: normalize webrtc.go to use receivers as pointers
Some receivers were being passed as values in webrtc.go, which is
incorrect since it was causing fields to be unmutated when they should,
and we'd lose reference to things like w.pc when closing the adapter
(i.e.: would cause PCs to stay up for a few seconds until they shut down
on state changes).

Normalize the struct so that pointers are preferred where applicable to
both fix aforementioned bugs and be a bit more efficient.