fix(livekit): deadlock on Close operation
There's a deadlock that might happen during the LK adapter's Close
operation. It's caused by a circular lock dependency between the
adapter's mutex and the recorder's mutex. The adapter's Close needs both
the adapter and recorder mutexes. This is also the case for some of the
recorder's ops (e.g.: pushing frames and needing to request a keyframe).
The recorder may indirectly wait for the adapter's mutex in one
of its routines (e.g.: pushing video, requesting keyframe) while its own
mutex is locked. This causes may cause the deadlock if an adapter Close
operation already had a hold of the adapter lock, but is waiting for
the recorder's.

Remove the uneeded adapter mutex Lock in its Close method to prevent the
aforementioned deadlock. Additionally, make a similar adjustment to the
KeyframeRequester setup where a similar scenario is likely to happen.

In the future, we need to review the mutexes and their lifecycles.
They're all over the place and a bunch of them could simply be replaced
by atomic vars.