fix(livekit): deadlock on GetStats call
There's another deadlock called by a circular dependency between
the LK adapter's and recorder's mutexes. This one is via the GetStats
method, which locks both the adapter and the recorder - while the
recorder also may do the same concurrently in a Push* call that triggers
a keyframe request.

Fix the deadlock by guaranteeing that the recorder is always called from
the adapter while the adapter is *unlocked*. Only the recorder may call
the hold the adapter and itself at the same time.