fix(livekit): add panic recovery when pushing packets
Add panic recovery to the LiveKit adapter's packet-pushing
goroutine. The expected behavior is for it to log the error, stop
processing and notify the affected recording session as failed so that
clients can handle that how they wish. This makes panics in that area
less harmful as they'll only affect one recording session directly.

Additionally:
- Add panic recovery to the WebM recorder's audio push function. It'll
  still be bubbled up to the adapter level, but we're trying to gather
  more info on a specific panic that's been happening.
- Remove a spurious keyframe request on audio frame writing failures.