fix(livekit): fix deadlock with jitterbuffer usage
A GetStats call locks both the adapter and the jitterbuffer
simultaneously. The later may also trigger an adapter lock via the
packet drop handling callback (which calls RequestKeyframe), hence
potentially causing a deadlock.

Refactor the adapter so that keyframe requests are made via a buffered
channel asynchronously when called via the packet drop handling
callback. This fixes the deadlock.

Additionally, review the rest of the lock calls in the adapter to
guarantee that no more self or external deadlocks are present.