refactor(livekit): optionally abort boot on LiveKit conn check
LiveKit is an optional component and, even though the health check
itself is configurable, aborting on boot is not. This is problematic
because BBB service hierarchy still does not treat LK as an obligatory
component, so there's no guarantee of boot order to begin with. Due to
this lack of hierarchy + LK being optional, aborting boot should be
optional as well.

Add a new livekit.healthCheck.abortBootOnFailure config flag. False by
default.