fix: rework start/stop command processing
The current command processing loop for a session currently blocks
request processing for start/stop calls. This may cause requests to clog
up in certain scenarios, including heartbeat checks. Those heartbeat
checks failing may cause clients to think the whole application is
unresponsive, which would not be the case. Additionally, this just
performs badly in any case other than the golden path.

Rework the request processing loop so that start/stop requests run in
separate goroutines for each session. The req/resp flow is largely the
same, but those two now go through a dedicated command queue for each
session that is processed by a goroutine. This should unblock the rest
of the requests/events.

Additional changes in this commit that are accessory to the above
changes:
  - Add a new Histogram metric: request_duration_seconds
  - Add a new counter metric: session_errors_total
    - Counts session command queue processing failures

Additional general improvements:
  - Extract stats types to a separate package so that it can be re-used
    across appstats, livekit and recorder without creating dependency
    loops. I had a hard time creating tests for the server.go pkg due to
    circular deps.