fix: don't emit stop event for sessions that have not started
The application currently emit stop events for sessions that fail during
their start routines. This is not the expected behavior of the API. It
should only return an StartRecordingResponse with a failure code and
reason. StopRecording events should only be fired for sessions that have
correctly started.

Guarantee stop events are not sent for not-yet-started sessions.
Add a regression test for this in the server package.