feat(livekit): add metrics for connect and subscribe request durations
We need to further clarify which parts of the startRecording pipeline
take the most time to set up in the higher percentiles. The connect and
subscribe procedures should be the longest ones, so some granular
metrics for them would be interesting.

Add specific metrics for LK's connect and subscribe procedures:
  - livekit_connect_duration_seconds
  - livekit_subscribe_duration_seconds